requirements.txt
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.59.0
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mistune==3.1.3
msgpack==1.1.1
mutagen==1.47.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numba==0.61.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.8
pooch==1.8.2
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==27.0.1
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.12.1
soupsieve==2.7
soxr==0.5.0.post1
stack-data==0.6.3
threadpoolctl==3.6.0
tinycss2==1.4.0
tk==0.1.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
youtube-search-python==1.6.6
yt-dlp==2025.1.12
